requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
dotenv-python==0.0.1
exceptiongroup==1.2.0
Faker==23.3.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
kafka-python==2.0.2
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.8.2
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
